Unnest, link to regex101 in comment, switch to ExternalLink
