+++ apps/hyperdrive-trading/src/ui/portfolio/shorts/ShortsContainer.tsx
@@ -1,23 +1,24 @@
 import { appConfig, HyperdriveConfig } from "@delvtech/hyperdrive-appconfig";
 import { Link } from "@tanstack/react-router";
 import { ReactElement } from "react";
+import { ExternalLink } from "src/ui/analytics/ExternalLink";
 import LoadingState from "src/ui/base/components/LoadingState";
 import { NonIdealState } from "src/ui/base/components/NonIdealState";
 import { ConnectWalletButton } from "src/ui/compliance/ConnectWallet";
-import { OpenLongsTableDesktop } from "src/ui/portfolio/longs/OpenLongsTable/OpenLongsTableDesktop";
-import { TotalOpenLongsValue } from "src/ui/portfolio/longs/TotalOpenLongsValue/TotalOpenLongsValue";
+import { PositionContainer } from "src/ui/portfolio/PositionContainer";
+import { PositionTableHeading } from "src/ui/portfolio/PositionTableHeading";
+import { OpenShortsTableDesktop } from "src/ui/portfolio/shorts/OpenShortsTable/OpenShortsTableDesktop";
+import { TotalOpenShortValue } from "src/ui/portfolio/shorts/OpenShortsTable/TotalOpenShortsValue";
 import {
-  OpenLongPositionsData,
-  usePortfolioLongsData,
-} from "src/ui/portfolio/longs/usePortfolioLongsData";
-import { PositionContainer } from "src/ui/portfolio/PositionContainer/PositionContainer";
-import { PositionTableHeading } from "src/ui/portfolio/PositionTableHeading/PositionTableHeading";
+  OpenShortPositionsData,
+  usePortfolioShortsData,
+} from "src/ui/portfolio/shorts/usePortfolioShortsData";
 import { useAccount } from "wagmi";
 
-export function OpenLongsContainer(): ReactElement {
+export function OpenShortsContainer(): ReactElement {
+  const { openShortPositions, openShortPositionsStatus } =
+    usePortfolioShortsData();
   const { address: account } = useAccount();
-  const { openLongPositions, openLongPositionsStatus } =
-    usePortfolioLongsData();
   if (!account) {
     return (
       <PositionContainer className="my-28">
@@ -29,34 +30,34 @@ export function OpenLongsContainer(): ReactElement {
     );
   }
 
-  if (openLongPositionsStatus === "loading") {
+  if (openShortPositionsStatus === "loading") {
     return (
       <PositionContainer>
         <LoadingState
-          heading="Loading your Longs..."
-          text="Searching for Long events, calculating current value and PnL..."
+          heading="Loading your Shorts..."
+          text="Searching for Shorts events, calculating current value and PnL..."
         />
       </PositionContainer>
     );
   }
 
-  if (openLongPositions?.every((position) => !position.openLongs.length)) {
+  if (
+    openShortPositions?.every((position) => position.openShorts.length === 0)
+  ) {
     return (
       <PositionContainer className="my-28">
         <NonIdealState
-          heading="No Longs found"
+          heading="No Shorts found"
           text={
             <p className="max-w-xl">
               Visit the{" "}
-              <a
-                className="daisy-link"
-                href="https://docs.hyperdrive.box/hyperdrive-overview/position-types/longs-fixed-rates"
-                rel="noopener noreferrer"
-                target="_blank"
+              <ExternalLink
+                newTab
+                href="https://docs.hyperdrive.box/hyperdrive-overview/position-types/shorts-variable-rates"
               >
                 documentation
-              </a>{" "}
-              or explore pools to open your first Long position.
+              </ExternalLink>{" "}
+              or explore pools to open your first Short position.
             </p>
           }
           action={
@@ -71,30 +72,31 @@ export function OpenLongsContainer(): ReactElement {
 
   return (
     <PositionContainer className="mt-10">
-      {openLongPositions &&
+      {openShortPositions &&
         appConfig.hyperdrives
           .filter((hyperdrive) => {
-            const openLongs = findOpenLongs(
-              openLongPositions,
+            const openShorts = findOpenShorts(
+              openShortPositions,
               hyperdrive,
-            )?.openLongs;
+            )?.openShorts;
             // Ensure this hyperdrive pool has open positions before rendering.
-            return openLongPositionsStatus === "success" && openLongs?.length;
+            return openShortPositionsStatus === "success" && openShorts?.length;
           })
           .map((hyperdrive) => {
-            const openLongs = findOpenLongs(
-              openLongPositions,
+            const openShorts = findOpenShorts(
+              openShortPositions,
               hyperdrive,
-            )?.openLongs;
+            )?.openShorts;
+
             return (
               <div className="flex flex-col gap-6" key={hyperdrive.address}>
                 <PositionTableHeading
                   hyperdrive={hyperdrive}
-                  rightElement={<TotalOpenLongsValue hyperdrive={hyperdrive} />}
+                  rightElement={<TotalOpenShortValue hyperdrive={hyperdrive} />}
                 />
-                <OpenLongsTableDesktop
+                <OpenShortsTableDesktop
                   hyperdrive={hyperdrive}
-                  openLongs={openLongs}
+                  openShorts={openShorts}
                 />
               </div>
             );
@@ -103,11 +105,11 @@ export function OpenLongsContainer(): ReactElement {
   );
 }
 
-function findOpenLongs(
-  openLongPositions: OpenLongPositionsData,
+function findOpenShorts(
+  openShortPositions: OpenShortPositionsData,
   hyperdrive: HyperdriveConfig,
 ) {
-  return openLongPositions.find(
+  return openShortPositions.find(
     (position) =>
       position.hyperdrive.address === hyperdrive.address &&
       position.hyperdrive.chainId === hyperdrive.chainId,
